Inventory/requirements.txt
asgiref==3.8.1
certifi==2021.10.8
charset-normalizer==2.0.12
cloudinary==1.29.0
dj-database-url==0.5.0
Django==5.1
django-cloudinary-storage==0.3.0
django-heroku==0.3.1
django-jazzmin==3.0.0
django-phonenumber-field==8.0.0
djangorestframework==3.13.1
idna==3.3
mysqlclient==2.2.4
phonenumbers==8.13.43
psycopg2==2.9.9
python-decouple==3.6
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
urllib3==1.26.8
whitenoise==6.0.0
